Rework SelectBiomePanel so it uses PagedSelector class.
